gh-95218: Move tests for importlib.resources into test_importlib.resources.
