+++ Lib/test/test_importlib/resources/data01/binary.file
