+++ Lib/test/test_importlib/resources/data01/subdirectory/binary.file
