+++ Lib/test/test_importlib/resources/data01/utf-16.file
