+++ Lib/test/test_importlib/resources/data01/utf-8.file
