+++ Lib/test/test_importlib/resources/data02/one/resource1.txt
