+++ Lib/test/test_importlib/resources/data02/two/resource2.txt
