+++ Lib/test/test_importlib/resources/namespacedata01/binary.file
