+++ Lib/test/test_importlib/resources/namespacedata01/utf-8.file
