+++ Lib/test/test_importlib/resources/test_files.py
@@ -4,7 +4,7 @@ import unittest
 from importlib import resources
 from importlib.abc import Traversable
 from . import data01
-from .resources import util
+from . import util
 
 
 class FilesTests:
