+++ Lib/test/test_importlib/resources/test_path.py
@@ -3,7 +3,7 @@ import unittest
 
 from importlib import resources
 from . import data01
-from .resources import util
+from . import util
 
 
 class CommonTests(util.CommonTests, unittest.TestCase):
