+++ Lib/test/test_importlib/resources/test_reader.py
