+++ Lib/test/test_importlib/resources/update-zips.py
