+++ Lib/test/test_importlib/resources/zipdata01/ziptestdata.zip
