+++ Lib/test/test_importlib/resources/zipdata02/ziptestdata.zip
