+++ Misc/NEWS.d/next/Tests/2022-07-24-17-24-42.gh-issue-95218.zfBLtu.rst
@@ -1,0 +1,1 @@
+Move tests for importlib.resources into test_importlib.resources.
